[REFACTOR] Guest 패키지 수정 (#153)
